// SysWas.Web/ClientApp/src/app/model/cep/cidade.ts
import { Estado } from './estado';

export class Cidade {


  id: number;
  nome: string;
  estadoId: number;
  estado: Estado;

}




// public long Id { get; set; }
      //  public string Nome { get; set; }
      //  public int EstadoId { get; set; }

      //  public int ThisEstadoId { get; set; }
      //  public virtual Estado ThisEstado { get; set; }